requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu113
# torch==1.11.0+cu113
# torchvision==0.12.0+cu113
# torchaudio==0.11.0

torch
torchvision
torchaudio

openmim
timm
opencv-python
termcolor
yacs
pyyaml
scipy
yapf==0.40.1    # for mmdet compatibility
pillow
rerun
tqdm
matplotlib
scikit-learn
